CloudInitWebCallback: arbitrary asynchronous hooks, part 1
As of know CloudInitWebCallback executes a number of actions with VM
info (hostname, ip, ssh key) in a dedicated thread. This actions are
"asynchronous" with respect to sending replies by http server to clients
(VMs which have "called home"). Initially the only activity was to update
~/.ssh/known_hosts with ssh keys of newly provisioned VMs, however there
are more things to do as of now:

- make ansible inventory
- print notifications

And I'd like to add even more, such as writing ~/.ssh/config alike file.
Thus _ssh_keys_updater looks cluttered, and the name is misleading.
Refactor the code so it's easy to add arbitrary tasks in the future